pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "pybind11>=2.10"]
build-backend = "setuptools.build_meta"

[project]
name = "bevel"
version = "0.2.0"
description = "A CLI-based 2D/3D game engine"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Bevel Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
]
dependencies = [
    "pyyaml>=6.0",
    "raylib>=5.0",
    "box2d>=2.3",
    "click>=8.0",
    "pybind11>=2.10",
    "watchdog>=3.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]
jolt = [
    "pybind11>=2.10",
]
hotreload = [
    "watchdog>=3.0",
]

[project.scripts]
bevel = "bevel.cli.main:cli"

[project.urls]
Homepage = "https://github.com/bevel-engine/bevel"
Documentation = "https://bevel-engine.github.io/bevel"
Repository = "https://github.com/bevel-engine/bevel"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
